Change tweet IDs to decimals

+++ db/migrate/20150403181555_create_queries.rb
@@ -1,7 +1,7 @@
 class CreateQueries < ActiveRecord::Migration
   def change
     create_table :queries do |t|
-      t.integer :last_tweet, null: false
+      t.decimal :last_tweet, null: false
       t.integer :count, null: false
       t.datetime :time, null: false
       t.integer :company_id, null: false
